src/requirements.txt
pyspark==3.5.1
snowflake-connector-python==3.10.0
requests
kafka-python==2.0.2
pandas
pyyaml
pytest
python-dotenv
# black
# flake8
# isort
# mypy
